pyproject.toml
[project]
name = "youtube-ai-v2"
version = "2.0.0"
description = "Minimal, resilient YouTube video generator from financial news"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "litellm>=1.0",
    "ffmpeg-python>=0.2",
    "pyttsx3>=2.90",
    "python-dotenv>=1.0",
    "pydub>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, no external deps)",
    "integration: Integration tests (mocked APIs)",
    "e2e: End-to-end tests (real APIs, slow)",
]
